upload section 2 & 3 (#2)
* upload section 2 & 3

Uploading the code from completion of section 2 & 3. Creating a branch pending no code interactions with future sections. Potential issue with the definition of the future value function, since it uses the same variables defined in the first method of calculation, as specified in the request. 
Could combine both loans into a single list of dictionaries and just use a loop to calculate future value of each loan.

* resolved inconsistencies, cleaned up code

Corrected the variables used by the two different sections such that the correct output is generated. Cleaned up and commented code.